+++ build.gradle
@@ -35,6 +35,7 @@ dependencies {
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.17.1'
     implementation 'org.commonmark:commonmark:0.21.0'
     implementation 'org.jsoup:jsoup:1.17.2'
+    implementation 'com.zaxxer:HikariCP:5.1.0'
 
     testImplementation("org.junit.jupiter:junit-jupiter-api:${junitVersion}")
     testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:${junitVersion}")
