+++ src/main/resources/hikari.properties
@@ -1,0 +1,5 @@
+dataSourceClassName=com.mysql.cj.jdbc.MysqlDataSource
+dataSource.url=jdbc:mysql://localhost:3306/blog?useSSL=false&useUnicode=true&allowPublicKeyRetrieval=true
+dataSource.user=root
+dataSource.password=1535
+maximumPoolSize=10
